requirements.txt
Django
djangorestframework
django-rest-knox
django-filter
psycopg2
drf-yasg
Pillow